Version 1.7.2 update!

+++ src/engine/forcefieldcurve.cpp
@@ -186,7 +186,7 @@ void ForceFieldCurve::_clearSubclassState() {
 void ForceFieldCurve::_addForceFieldToGridMT(MACVelocityField &fieldGrid, int dir) {
     int U = 0; int V = 1; int W = 2;
 
-    int gridsize = 0;
+    size_t gridsize = 0;
     if (dir == U) {
         gridsize = (_isize + 1) * _jsize * _ksize;
     } else if (dir == V) {
@@ -195,7 +195,7 @@ void ForceFieldCurve::_addForceFieldToGridMT(MACVelocityField &fieldGrid, int di
         gridsize = _isize * _jsize * (_ksize + 1);
     }
 
-    int numCPU = ThreadUtils::getMaxThreadCount();
+    size_t numCPU = ThreadUtils::getMaxThreadCount();
     int numthreads = (int)fmin(numCPU, gridsize);
     std::vector<std::thread> threads(numthreads);
     std::vector<int> intervals = ThreadUtils::splitRangeIntoIntervals(0, gridsize, numthreads);
